feat: separate run and paragraph
